spectator mode && fixed patrick almost underflow
My buddy Patrick discovered hitting backspace at the username prompt caused a
DoS.  It would of been an underflow if len was of int type.  That would of been
a sweet vulnerability because it would let an attacker write ascii chars to the
malloc header BUT because len is a size_t you end up putting a NULL byte
somewhere you shouldn't, you have little control where, then immediately
hitting a FATAL when strlen(buff) != len.