telnet are you there implemented and parse lists.
Players who have not done anything for a while will be asked if they are still
there using the telnet protocol, if they don't respond appropriately they are
kicked.

Config file can have a list of things in it, currently only works for the bad
usernames.  This should work better also, the list should be created in an
order.