fixed overflow, started signal handling
There was a heap overflow that likely was not exploitable except for a DoS. When
the server is configured to have more then 5 players the SERVER.players array
was not reallocated to reflect any change in size. This meant if you ran this
server and accepted 10 players, 5 * (sizeof(void *)) bytes could be overwritten
just by someone connecting to the host 10 times.

The bytes were overwritten with the addresses in the heap, and NULL is written
when they are done. Not a lot of room to play around.

Another security fix is removing the ability to designate a config file from
command line. This is silly and needs to be updated much more fully. Hardly a
fix at all.

Lastly, basic signal handling was added... There is still much to do and memory
leaks are still plentiful.